What the fuck Noc. Magic Rework (#7351)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Thanks to Azure for the idea and framework for some of the new spells
Someone tell Noc to stop reading TTRPG manually please.
Redoes the magic system to one thats easier to maintain and balance and
pave the way for future class types.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Magic sucks, its an established fact. The Mages are nerds, the
attunements are wack. How are we gonna fix this? We are just gonna kill
em!
This basically nukes everything complex about the old magic system and
redoes it in a way thats easier to digest.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
del: player attunements
del: spell attunements
add: spell forms and techniques, spells are now split into a form and an
optional technique
add: bookcrafting system, now books can take more then a catalyst and
doing so provides buffs for all spells
refactor: spells learned from your spellbook are now only kept if you
are holding the spellbook.
add: spellsword and spellfist spells
add: 100 new impact spells
add: 40 new cantrip spells
refactor: lightning spells no longer hard stun constantly
add: arcyne necromancy
refactor: summoning, it now has a downside where when your summons die
you experience backlash preventing resummon
add: sorcerer framework, unlike mages sorcerers cannot learn spells from
a spellbook in exchange for more innate points
add: spellbook granted spells now have a limited number of charges, to
replenish you need to use the rune of recharge.
del: attunement matrix
refactor: spell cost, spell cast speed, and spell strength are now
public information on your equipment so its not impossible to tell what
is giving you buffs, and is no longer mutated by the spell you are
casting
qol: you can now examine spells with shift click to get name and desc
change: sending a spell to the end of the list is now alt-click instead
of shift-click
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

+++ code/modules/mapping/reader.dm
@@ -952,6 +952,8 @@ GLOBAL_LIST_EMPTY(map_model_default)
 			for(var/atom/turf_atom as anything in crds.GetAllTurfStrictContents())
 				if(isobserver(turf_atom))
 					continue
+				if(istype(turf_atom, /obj/effect/ebeam))
+					continue
 				qdel(turf_atom, force = TRUE)
 
 		// Note: we make the assertion that the last path WILL be a turf. if it isn't, this will fail.
